Introduce new module for HeadLogic types
